Created new task from E-Olimp

+++ src/E_Olymp/Task1602.java
@@ -2,21 +2,22 @@ package E_Olymp;
 
 import java.util.Scanner;
 
+//100
 public class Task1602 {
     public static void main(String[] args) {
         Scanner scanner=new Scanner(System.in);
-        int a=scanner.nextInt();
-        int b=scanner.nextInt();
+        long a=scanner.nextLong();
+        long b=scanner.nextLong();
         System.out.println(ekob(a,b));
 
     }
 
-     static int ebob(int a ,int b){
+     static long ebob(long a ,long b){
         if(a%b==0) return b;
         return ebob(b, a%b);
     }
 
-      static int ekob(int a,int b){
+      static long ekob(long a,long b){
         if (a%b==0) return a;
         return a*b / ebob(a,b);
     }
